+++ imap_tools/__init__.py
@@ -4,4 +4,4 @@ from .message import MailMessage, MailAttachment, MailMessageFlags
 from .folder import MailBoxFolderManager, MailBoxFolderStatusOptions
 from .errors import *
 
-__version__ = '0.38.1'
+__version__ = '0.39.0'